fix(work-cn): 积分浮点解析、点×最小化到托盘、trae_cn 平台兼容、覆盖安装杀旧进程
- 积分改 f64 解析（实测用量带小数 864.63，整数解析读成 0）

- 点 × 改为最小化到托盘并销毁 WebView 释放内存，Ask 兼容为 Minimize

- is_work_cn_account_kind 宽松判定 trae_cn/trae_solo_cn，导入强制规范化平台，导入链路合并设备快照防刷新覆盖

- NSIS 安装前终止旧进程，支持新版本直接覆盖安装

+++ src-tauri/src/models/work_cn.rs
@@ -93,9 +93,11 @@ pub struct WorkCnSwitchResult {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct WorkCnCreditsSummary {
-    pub total: Option<i64>,
-    pub used: i64,
-    pub remaining: Option<i64>,
+    /// 浮点：真实接口的 `usage.credits_amount` 带小数（实测 864.63），
+    /// 整数解析会把用量读成 0，导致“剩余积分”虚高。
+    pub total: Option<f64>,
+    pub used: f64,
+    pub remaining: Option<f64>,
     pub unlimited: bool,
     pub updated_at: i64,
 }
